requirements.txt
# Financial Platform Backend Requirements

fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
sqlalchemy==2.0.43
alembic==1.16.5
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.3.0
pandas==2.3.2
numpy==2.3.3
scipy==1.16.1
requests==2.32.5
aiohttp==3.12.15
websockets==15.0.1
yfinance==0.2.65
python-dotenv==1.1.1
PyPortfolioOpt==1.5.6
backtesting==0.6.5
hmmlearn==0.3.3
fbm==0.3.0
anthropic==0.66.0
